Implement onboarding components and restructure navigation for Fyxlife app
