+++ src/components/index.ts
@@ -1,11 +1,11 @@
 // Container Components
-export { default as WelcomeContainer } from './containers/WelcomeContainer';
-export { default as UserInfoContainer } from './containers/UserInfoContainer';
-export { default as ConfirmationContainer } from './containers/ConfirmationContainer';
-export { default as MainAppContainer } from './containers/MainAppContainer';
+export { default as WelcomeContainer } from './containers/onboarding/WelcomeContainer';
+export { default as UserInfoContainer } from './containers/onboarding/UserInfoContainer';
+export { default as ConfirmationContainer } from './containers/onboarding/ConfirmationContainer';
+export { default as MainAppContainer } from './containers/onboarding/MainAppContainer';
 
 // Presentation Components
-export { default as WelcomePresentation } from './presentation/WelcomePresentation';
-export { default as UserInfoPresentation } from './presentation/UserInfoPresentation';
-export { default as ConfirmationPresentation } from './presentation/ConfirmationPresentation';
-export { default as MainAppPresentation } from './presentation/MainAppPresentation';
+export { default as WelcomePresentation } from './presentation/onboarding/WelcomePresentation';
+export { default as UserInfoPresentation } from './presentation/onboarding/UserInfoPresentation';
+export { default as ConfirmationPresentation } from './presentation/onboarding/ConfirmationPresentation';
+export { default as MainAppPresentation } from './presentation/onboarding/MainAppPresentation';
