+++ src/navigation/screens/onboarding/ConfirmationScreen.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ConfirmationContainer from '../../components/containers/ConfirmationContainer';
+import ConfirmationContainer from '../../../components/containers/onboarding/ConfirmationContainer';
 
 export default function ConfirmationScreen() {
   return <ConfirmationContainer />;
