+++ src/navigation/screens/onboarding/MainAppScreen.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import MainAppContainer from '../../components/containers/MainAppContainer';
+import MainAppContainer from '../../../components/containers/onboarding/MainAppContainer';
 
 export default function MainAppScreen() {
   return <MainAppContainer />;
