+++ src/navigation/screens/onboarding/UserInfoScreen.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import UserInfoContainer from '../../components/containers/UserInfoContainer';
+import UserInfoContainer from '../../../components/containers/onboarding/UserInfoContainer';
 
 export default function UserInfoScreen() {
   return <UserInfoContainer />;
