+++ src/navigation/screens/onboarding/WelcomeScreen.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import WelcomeContainer from '../../components/containers/WelcomeContainer';
+import WelcomeContainer from '../../../components/containers/onboarding/WelcomeContainer';
 
 export default function WelcomeScreen() {
   return <WelcomeContainer />;
